% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDOM440.R
\name{Abs.val}
\alias{Abs.val}
\title{Extract CDOM value at chosen nm}
\usage{
Abs.val(data.file = "data", FileSelect = F, skip = 1, nm = 440)
}
\arguments{
\item{FileSelect}{is a logical argument for the number of files to load. Default is FALSE, meaning that the function}

\item{skip}{is the number of lines to skip before getting the header in CDOM files}

\item{nm1}{is the desired wavelength. Default is 440}
}
\description{
This function reads any number of CDOM files and select an absorbance value.
}
