% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIX.R
\name{BIX}
\alias{BIX}
\title{Calculate BIX}
\usage{
BIX(cube)
}
\description{
This function works with the output of PARAFAC.cube.design and calculate BIX.
}
\references{
Biological index (BIX) from Huguet et al. 2009
}
