% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDOMFileCorrection.R
\name{CDOMFileCorrection}
\alias{CDOMFileCorrection}
\title{This function fills gap in missing nanometers between CDOM and FDOM matching file}
\usage{
CDOMFileCorrection(CDOM, low, high, lowfrom = 0, lowto = 0,
  highfrom = 0, highto = 0)
}
\arguments{
\item{CDOM}{is the CDOM file to correct}

\item{low}{is a logical parameter to let the function add short wavelengths.}

\item{high}{is a logical parameter to let the function add long wavelengths.}

\item{lowfrom}{is the start of the short wavelength}

\item{lowto}{is the end of the short wavelength}

\item{highfrom}{is the start of the long wavelength}

\item{highto}{is the end of the long wavelength}
}
\description{
If CDOM files are short in absorbance at low or high wavelength compared to EEMs
this function will create dummy values to fill the gap and let the function apply the correction
These wavelength must be removed from the EEMs
}
