% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDOMSpectreOverlay.R
\name{CDOMOverlay}
\alias{CDOMOverlay}
\title{Superimpose all absorbance scan}
\usage{
CDOMOverlay(pdf = F)
}
\arguments{
\item{pdf}{is a logical paramter to export curves in a pdf file. Default is F}
}
\description{
This function reads all files in /CDOM and plot them.
}
