% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaEEM.R
\name{DeltaEEM}
\alias{DeltaEEM}
\title{Substract a corrected EEM from another}
\usage{
DeltaEEM(cube, diminuende, diminuteur, exact_ende = T, exact_teur = T,
  zlim = c(-1, 1))
}
\arguments{
\item{cube}{is the output of the main function PARAFAC.cube.design()}

\item{diminuende}{is the sample you want to subtract from (Ex. a - b = c, a is the diminuende)}

\item{diminuteur}{is the sample you want to subtract (Ex. a - b = c, b is the diminuteur)}

\item{zlim}{is the intensity to plot the EEMs. Negative and values outside of scale bar will be white
If paramter is set to FALSE, LTM = LTMGose50x. Beware of potential confusion if names are similar!}

\item{Exact_ende}{is a logical parameter for diminuende name. Default is TRUE (Ex. Bunker != BunkerIPA50x)
If paramter is set to FALSE, Bunker = BUNKERIPA50x. Beware of potential confusion if names are similar!}

\item{Exact_teur}{is a logical parameter for diminuteur name. Default is TRUE (Ex. LTM != LTMGose50x)}
}
\description{
This function select 2 EEMs based on their name and subtract one from another
}
