% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DilutionFactor.R
\name{DilutionFactor}
\alias{DilutionFactor}
\title{Multiply EEMs by their dilution factor}
\usage{
DilutionFactor(cube, vector.out = F)
}
\arguments{
\item{cube}{is the output of PARAFAC.cube.design}

\item{vector.out}{is a logical parameter to return dilution factor vector. Default is FALSE, returning corrected cube}
}
\description{
This function recognize dilution factor in CDOM file name if written properly
The output returns the data cube corrected for dilution or a vector with dilution factor
Name example: SampleName_15x.csv
}
