% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportEEM.R
\name{ExportEEM}
\alias{ExportEEM}
\title{This function export your data into 1 single txt files that will be incorporated in Matlab for the PARAFAC}
\usage{
ExportEEM(cube, cut = c(0, 0, 0))
}
\arguments{
\item{cube}{is the output of PARAFAC.cube.design()}

\item{cut}{is the number of excitation, emission and sample to remove from cube, respectively. Default is c(0,0,0)}
}
\description{
This function export in text file the output of PARAFAC.cube.design()
It is possible to cut the cube if all excitation, emission or samples are not required
}
