% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDOMIndices.R
\name{FDOMIndices}
\alias{FDOMIndices}
\title{Generic functions for commonly used FDOM indices}
\usage{
FDOMIndices(cube, cor = T, rel = F)
}
\arguments{
\item{cube}{is the output of PARAFAC.cube.design()}

\item{cor}{is a logical argument for instruments corrections applied to EEMs. Default is TRUE}

\item{rel}{is a logical argument to scale the values of HIX}
}
\description{
This is a warp up function to calculate freshness index, FI, BIX and HIX based on the output of PARAFAC.cube.design()
}
\references{
Fluorescence index (FI) from McKnight et al (2001) and Cory and McKnight (2005)

Freshness index from Parlanti et al. (2000)

Humification index (HIX) from Ohno et al (2002)

Biological index (BIX) from Huguet et al (2009)
}
