% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FI.R
\name{FI}
\alias{FI}
\title{Calculate FI}
\usage{
FI(cube, cor = T)
}
\description{
This function works with the output of PARAFAC.cube.design and calculate FI.
}
\references{
Fluorescence index (FI) from McKnight et al (2001) and Cory and McKnight (2005)
}
