% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderCreation.R
\name{FolderCreation}
\alias{FolderCreation}
\title{Creates data folder, subfolder and R project}
\usage{
FolderCreation(path.to.folder, WorkingDirectory, subfiles = c("CDOM", "FDOM",
  "nano"))
}
\arguments{
\item{path.to.folder}{is the path where the working directory is}

\item{WorkingDirectory}{is the name of the working directory}

\item{subfiles}{is the list of folder where to put cdom, fdom and nano water samples}
}
\description{
This function creates folders and .Rproj file require by the various functions to work.
These steps can be done by hand instead.
}
