% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreshIndex.R
\name{FreshIndex}
\alias{FreshIndex}
\title{Calculate Freshness index}
\usage{
FreshIndex(cube)
}
\description{
This function works with the output of PARAFAC.cube.design and calculate BIX.
}
\references{
Freshness index from Parlanti et al. 2000
}
