% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIX.R
\name{HIX}
\alias{HIX}
\title{Calculate HIX}
\usage{
HIX(cube, rel = FALSE)
}
\description{
This function works with the output of PARAFAC.cube.design and calculate HIX.
}
\references{
Humification index (HIX) from Ohno et al (2002)
}
