% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSD.R
\name{NanoSD}
\alias{NanoSD}
\title{This function calculates and plots standard deviation of nano files}
\usage{
NanoSD(excitation = c(220, 450, 5), emission = c(230, 600, 2), EMCOL = F,
  split = "_", data.file = "data", fluorometer = "Cary Eclipse", EEMskip,
  sd = T, zlim = c(0, 1))
}
\arguments{
\item{excitation}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(220,450,5)}

\item{emission}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(230,600,2)}

\item{EMCOL}{is a logical parameter indicating whether or not the emission are
stored as column in the csv file. Default is FALSE.}

\item{split}{is the symbol used to separate each EEM when using a multicell holder}

\item{data.file}{is the name of the main folder where CDOM, FDOM and nano folders are}

\item{fluorometer}{is a paramater for the fluorometer model. Default is "Cary Eclipse".
Other model supported: "Shimadzu"}

\item{EEMskip}{is a parameter to skip lines in EEM file before data. Default is 1}

\item{sd}{is a logical parameter to calculate standard deviation with the NanoMean function. Default is True}

\item{zlim}{is the intensity to plot the EEMs. Negative and values outside of scale bar will be white}
}
\description{
This function uses NanoMean to compute the standrad deviation to make sure that all nano
files are similare
}
