% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAFAC.cube.design.R
\name{PARAFAC.cube.design}
\alias{PARAFAC.cube.design}
\title{Combine EEMs produce by the Cary Eclipse fluorometer into a cube
ready for PARAFAC analysis in Matlab}
\usage{
PARAFAC.cube.design(data.file = "data", excitation = c(220, 450, 5),
  emission = c(230, 600, 2), EMCOL = F, Subtract.Blank = T, RU = T,
  rm.corner = T, EmEx.cor = T, Inner = T, pathlength = 1,
  split = "_", skip = 1, dot.number = 1, NonNegativity = T,
  fluorometer = "Cary Eclipse", EEMskip = 1, rm.scat = T,
  zlim = c(0, 1), graph = F, RspectroAbs = F)
}
\arguments{
\item{data.file}{is the name of the main folder where CDOM, FDOM and nano folders are}

\item{excitation}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(220,450,5)}

\item{emission}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(230,600,2)}

\item{EMCOL}{is a logical parameter indicating whether or not the emission are
stored as column in the csv file. Default is FALSE.}

\item{RU}{is a logical parameter to transform fluorescence intensities into Raman Unit at Ex = 350 nm.
Default is TRUE.}

\item{rm.corner}{is a logical parameter to set the bottom right corner values to 0}

\item{EmEx.cor}{is a logical parameter to correct EEMs for emission and excitation corrections.
Default is True. Emission and excitation file must be numerics only stored in csv file.}

\item{Inner}{is a logical parameter to apply inner filter effect correction. Default is True}

\item{pathlength}{is a real number indicating the length of the optical path in the spectrophotometer in cm.
Default is 1}

\item{split}{is the symbol used to separate each EEM when using a multicell holder}

\item{skip}{is a parameter to determine how many lines will be skiped before the header in the absorbance files}

\item{dot.number}{is the number of "." in the name of your EEMs file. This number includes the "." in ".csv"}

\item{NonNegativity}{is a logical parameter to transform all negative fluorescence values into 0. Default is TRUE}

\item{fluorometer}{is a paramater for the fluorometer model. Default is "Cary Eclipse".
Other model supported: "Shimadzu"}

\item{EEMskip}{is a parameter to skip lines in EEM file before data. Default is 1}

\item{rm.scat}{is a logical parameter to remove Reyleigh and Raman scatterings.}

\item{zlim}{is the intensity to plot the EEMs. Negative and values outside of scale bar will be white}

\item{graph}{is a logical parameter to plot the Raman integration to transform data into RU. Default is false}

\item{RspectroAbs}{is a logical parameter for the inner filter correction. When TRUE, CDOM data have been first treated by the 
RspectroAbs package developed by Simon Bélanger. Default is FASLE}
}
\description{
This function reads any number of EEMs and produce a cube of EEMs.
It enables the standardization into Raman Unit and instrument's corrections.
}
