% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eem.go.R
\name{plot.EEM.go}
\alias{plot.EEM.go}
\title{Subtract and plot fluorescence Excitation-Emission Matrix (EEM)}
\usage{
\method{plot}{EEM.go}(path, zlim = c(0, 10), SUBTRACT.BLANK = TRUE,
  PLOT.RAMAN = TRUE, PLOT.BLANK = TRUE, excitation = c(220, 450, 5),
  emission = c(230, 600, 2), EMCOL = FALSE, samplepercsv = 1, RU = T)
}
\arguments{
\item{path}{Full path of the working directory (can be called using getwd())}

\item{zlim}{is the limits of the fluoresence scale (z axis) in raw units. Default is c(0, 10) for clear oceanic waters.}

\item{SUBTRACT.BLANK}{is a logical parameter to indicate if a blank EMM is to be subtract from the sample EEM.
Default is TRUE and the user is prompted to first select the blank file.}

\item{PLOT.RAMAN}{is a logical parameter indicating
whether or not the raman peak is ploted. Default is TRUE.}

\item{PLOT.BLANK}{is a logical parameter indicating whether the blank EEMs is ploted or not. Default is TRUE.}

\item{excitation}{is a vector of three variables of the scanning setup (min, max, interval).
Default is c(220, 450, 5)}

\item{emission}{is a vector of three variables of the scanning setup (min, max, interval).
Default is c(230, 600, 2)}

\item{EMCOL}{is a logical parameter indicating whether or not the emission are
stored as column in the csv file. Default is FALSE.}

\item{samplepercsv}{is a parameter which indicates the number of sample in the csv file coming from the fluorometer.}

\item{RU}{is a logical parameter to transform fluorescence intensities into Raman Unit at Ex = 350 nm.
Default is TRUE}
}
\value{
Returns the EEM from the sample
}
\description{
A simple function to visualize EEMs obtained using a Varian Cary Eclipse fluorometer in csv format
(could evolved to include other instrument). The user is prompted to select one or many files
}
\details{

}
\examples{

plot.EEM.go(getwd(), zlim = c(0, 20), SUBTRACT.BLANK = FALSE)

}
\seealso{
\code{\link{read.EEM}}
}
\author{
Simon Bélanger & Richard LaBrie
}
