% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllEEM.R
\name{PlotAllEEM}
\alias{PlotAllEEM}
\title{This function plot all corrected EEMs}
\usage{
PlotAllEEM(cube, zlim = c(0, 1), ExportPDF = T)
}
\arguments{
\item{cube}{is the ouput of PARAFAC.cube.design}

\item{zlim}{is the intensity to plot the EEMs. Negative and values outside of scale bar will be white}

\item{ExportPDF}{is a logical paramater to export plots in a pdf file. Default is TRUE}
}
\description{
This function uses the output of PARAFAC.cube.design to plot all corrected EEMss at once
The scale bar is the same among all samples. It also generates a pdf file containing all plots
}
