% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EEM.R
\name{read.EEM}
\alias{read.EEM}
\title{Reads CSV file of EEMs produced by the Cary Eclipe fluorometer}
\usage{
read.EEM(filename, excitation = c(220, 450, 5), emission = c(230, 600, 2),
  EMCOL = FALSE, samplepercsv = 1, split = "_", dot.number = 1)
}
\arguments{
\item{excitation}{is a vector of three variables of the scanning setup (min, max, interval).
Default is c(220, 450, 5)}

\item{emission}{is a vector of three variables of the scanning setup (min, max, interval).
Default is c(230, 600, 2)}

\item{EMCOL}{is a logical parameter indicating whether or not the emission are
stored as column in the csv file. Default is FALSE.}

\item{samplepercsv}{is a parameter which indicates the number of sample}

\item{split}{is a parameter to define the symbol between your samples name if samplepercsv > 1.
Default is "_"}
}
\description{
This function reads a CSV file produced by the Cary Eclipe fluorometer
}
\seealso{
\code{\link{plot.EEM.go}}
}
