% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RspectroAbsConvert.R
\name{RspectroAbsConvert}
\alias{RspectroAbsConvert}
\title{RspectroAbs converter to csv files}
\usage{
RspectroAbsConvert(Rdatapath = "./data/Rdata/",
  CDOMpath = "./data/CDOM/")
}
\arguments{
\item{Rdatapath}{is the path to you Rdata folder created by run.process.ag.batch(). Default is "./data/Rdata/"}

\item{CDOMpath}{is the path to your CDOM folder used for this package. Default is "./data/CDOM/"}
}
\description{
Transforms Rdata files created by RspectroAbs package developed 
by Simon Bélanger into csv files usable by this package
}
