% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RspectroAbsIndices.R
\name{RspectroAbsIndices}
\alias{RspectroAbsIndices}
\title{Summarize RspectroAbs CDOM parameters}
\usage{
RspectroAbsIndices(RDatapath = "./data/Rdata/", export = F)
}
\arguments{
\item{export}{is a logical paramater to export a matrix with all CDOM indices. Default is FALSE}

\item{Rdatapath}{is the path to you Rdata folder created by run.process.ag.batch(). Default is "./data/Rdata/"}
}
\description{
Summarizes all CDOM indices calculated with RspectroAbs package developed 
by Simon B攼㸹langer and export them in csv
}
