% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralSlope.R
\name{SpectralSlope}
\alias{SpectralSlope}
\title{Generate the spectral slope for CDOM samples}
\usage{
SpectralSlope(data.file = "data", FileSelect = F, wl0 = 375, From = 275,
  To = 650, By = 1, skip = 1)
}
\arguments{
\item{FileSelect}{is a logical parameter to allow you which file to load. Default is FALSE, meaning that all files in /CDOM
will be loaded}

\item{wl0}{is the reference wavelength, default is 375 as in Stedmon, Markager and Kaas 2000}

\item{To}{is the wavelenth at which you want to end calculating the exponential fit}

\item{By}{is the step between each wavelength}

\item{skip}{is a parameter to determine how many lines will be skiped before the header in the absorbance files}

\item{FROM}{is the wavelenth at which you want to start calculating the exponential fit}
}
\description{
This function reads any number of CDOM files and calculate their spectral slope.
You have to manually load DOC file for SUVA
Wavelengths chosen based on Helms 2008
Calculate spectral slope with a reference wavelength
The model use a nonlinear least square fitting routine
The equation is abs = abs0 * exp(-S(wavelength - ref.wavelength)) + K as in Markager & Vincent 2000
where abs is the absorption coef, abs0 is the reference absorption coef at wavelenth and ref.wavelength
S is the slope and K is an additional background parameter to allow baseline to shift
}
