% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrSpectralSlope.R
\name{Sr.spectralslope}
\alias{Sr.spectralslope}
\title{Calculate Sr for CDOM samples}
\usage{
Sr.spectralslope(
  data,
  wl1.1 = 275,
  wl1.2 = 295,
  wl2.1 = 350,
  wl2.2 = 400,
  wl0 = 375
)
}
\description{
This function reads any number of CDOM files and calculate their Sr.
}
