% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUVA.R
\name{SUVA}
\alias{SUVA}
\title{Calculate SUVA indice}
\usage{
SUVA(data.file = "data", FileSelect = F, DOC, wl = 254, unit = "mg/L",
  skip = 1, name = "SUVAMatches")
}
\arguments{
\item{FileSelect}{is a logical parameter to allow you which file to load. Default is FALSE, meaning that all files in /CDOM
will be loaded}
}
\description{
This function reads any number of CDOM files and calculate SUVA.
You have to manually load DOC file for SUVA
Wavelengths chosen based on Helms 2008
Calculate spectral slope with a reference wavelength
The model use a nonlinear least square fitting routine
The equation is abs = abs0 * exp(-S(wavelength - ref.wavelength)) + K as in Markager & Vincent 2000
where abs is the absorption coef, abs0 is the reference absorption coef at wavelenth and ref.wavelength
S is the slope and K is an additional background parameter to allow baseline to shift
}
