% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDOMIndices.R
\name{FI}
\alias{FI}
\title{Generic functions for commonly used FDOM indices}
\usage{
FI(cube, cor = T)
}
\arguments{
\item{cube}{is the output of PARAFAC.cube.design()}

\item{cor}{is a logical parameter for instruments corrections applied to EEMs. Default is TRUE}
}
\description{
These functions are designed to work with the output of PARAFAC.cube.design and calculate basic FDOM indices.
}
