% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderCreation.R
\name{FolderCreation}
\alias{FolderCreation}
\title{Creates data folder, subfolder and R project}
\usage{
FolderCreation(
  WDpath = ".",
  folder = "data",
  subfiles = c("CDOM", "FDOM", "nano"),
  example = TRUE
)
}
\arguments{
\item{WDpath}{the path to the working directory}

\item{folder}{is the name of the folder containing subfiles parameter. Default is "data"}

\item{subfiles}{is the list of folder where to put cdom, fdom and nano water samples}

\item{example}{TRUE if you wich to donwload exemple files to test the package}
}
\description{
This function creates folders and .Rproj file require by the various functions to work.
These steps can be done by hand instead.
}
