% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAFAC.cube.design.R
\name{PARAFAC.cube.design}
\alias{PARAFAC.cube.design}
\title{Combine EEMs produce by the Cary Eclipse fluorometer into a cube
ready for PARAFAC analysis in Matlab}
\usage{
PARAFAC.cube.design(path = getwd(), excitation = c(220, 450, 5),
  emission = c(230, 600, 2), EMCOL = F, Subtract.Blank = T, RU = T,
  rm.corner = T, EmEx.cor = T, Inner = T, pathlength = 1, split = "_",
  skip = 1, dot.number = 1)
}
\arguments{
\item{excitation}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(220,450,5)}

\item{emission}{is a vector of three variables of the scanning setup (min,max,interval).
Default is c(230,600,2)}

\item{EMCOL}{is a logical parameter indicating whether or not the emission are
stored as column in the csv file. Default is FALSE.}

\item{RU}{is a logical parameter to transform fluorescence intensities into Raman Unit at Ex = 350 nm.
Default is TRUE.}

\item{EmEx.cor}{is a logical parameter to correct EEMs for emission and excitation corrections.
Default is True. Emission and excitation file must be numerics only stored in csv file.}

\item{skip}{is a parameter to determine how many lines will be skiped before the header in the absorbance files}

\item{dot.number}{is the number of "." in the name of your EEMs file. This number includes the "." in ".csv"}
}
\description{
This function reads any number of EEMs and produce a cube of EEMs.
It enables the standardization into Raman Unit and instrument's corrections.
}
